Fix responsibilty issue between the two files
